+++ src/ReceiptList.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import Receipt from './Receipt';
+
+const ReceiptList = ({ receipts }) =>
+  receipts.map((receipt) => <Receipt key={receipt.id} receipt={receipt} />);
+
+export default ReceiptList;
