+++ src/Receipts.jsx
@@ -1,21 +1,8 @@
-import {
-  SimpleGrid,
-  Card,
-  CardBody,
-  Heading,
-  Text,
-  CardHeader,
-  CardFooter,
-  Button,
-  Flex,
-  Box,
-  Stack,
-  StackDivider,
-  Spinner,
-} from '@chakra-ui/react';
+import { SimpleGrid, Box, Spinner } from '@chakra-ui/react';
 import { Link } from 'react-router-dom';
+
+import ReceiptList from './ReceiptList';
 import useFetch from './useFetch';
-import { useHistory } from 'react-router-dom/cjs/react-router-dom.min';
 
 export const Receipts = () => {
   const {
@@ -23,8 +10,6 @@ export const Receipts = () => {
     isLoading,
     error,
   } = useFetch('http://localhost:8000/receipts');
-
-  const history = useHistory();
 
   if (isLoading) {
     return <Spinner />;
@@ -36,42 +21,7 @@ export const Receipts = () => {
 
   return (
     <SimpleGrid columns={4} spacing={5}>
-      {receipts.map((receipt) => (
-        <Card key={receipt.id}>
-          <CardHeader>
-            <Heading size="md">{receipt.name}</Heading>
-          </CardHeader>
-          <CardBody>
-            <Stack divider={<StackDivider />}>
-              <Box>
-                <Heading size="xs" textTransform="uppercase">
-                  Total
-                </Heading>
-                <Text pt="2" fontSize="sm">
-                  $
-                  {receipt.items.reduce(
-                    (total, item) => (total += item.price * item.quantity),
-                    0
-                  )}
-                </Text>
-              </Box>
-              <Box>
-                <Heading size="xs" textTransform="uppercase">
-                  Created At
-                </Heading>
-                <Text pt="2" fontSize="sm">
-                  {new Date(receipt.created_at).toUTCString()}
-                </Text>
-              </Box>
-            </Stack>
-          </CardBody>
-          <CardFooter>
-            <Button onClick={() => history.push(`/receipts/${receipt.id}`)}>
-              View
-            </Button>
-          </CardFooter>
-        </Card>
-      ))}
+      <ReceiptList receipts={receipts} />
       <Box
         border="1px"
         borderColor="gray.100"
